# openapi.yaml
openapi: 3.1.0
info:
  title: DevMate Core API Specification
  description: |
    Production-ready REST and RPC API contract for the DevMate Telegram-based Personal Operating System platform.
    Defines schemas, endpoints, error response wrappers, webhooks, and callbacks for delivery adapters and future clients.
  version: 1.0.0
  termsOfService: https://devmate.internal/terms
  contact:
    name: DevMate API Platform Team
    email: platform-api@devmate.internal
    url: https://devmate.internal/support
  license:
    name: Proprietary License
    url: https://devmate.internal/license
...
servers:
  - url: https://api.devmate.internal/v1
    description: Production Server Gateway
  - url: https://staging-api.devmate.internal/v1
    description: Staging Environment Server
  - url: https://dev-api.devmate.internal/v1
    description: Development Integration Server
  - url: http://localhost:8080/v1
    description: Local Developer Container Server
security:
  - BearerAuth: []
paths:
...
  /users/me:
    get:
      summary: Retrieve current user profile
      description: Gets the authenticated user's core details.
      operationId: getCurrentUser
      tags:
        - Users
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/ServerError'
  /users/me/profile:
    patch:
      summary: Update profile details
      description: Modifies names and username fields for the current user.
      operationId: updateProfile
      tags:
        - Profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /users/me/preferences:
    patch:
      summary: Update preferences configuration
      description: Update language, base currency, and timezones preferences.
      operationId: updatePreferences
      tags:
        - Settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePreferencesRequest'
      responses:
        '200':
          description: Preferences updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreferencesResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /settings:
    get:
      summary: Get user settings
      description: Retrieve global flags, notification switches, and quiet hour schedules.
      operationId: getSettings
      tags:
        - Settings
      responses:
        '200':
          description: Settings retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    patch:
      summary: Update user settings
      description: Modifies notification preferences and schedules.
      operationId: updateSettings
      tags:
        - Settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSettingsRequest'
      responses:
        '200':
          description: Settings updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /todos:
    get:
      summary: List todo items
      description: Retrieve user checklist tasks. Supports status filtering and paginated pages.
      operationId: listTodos
      tags:
        - Todos
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/TodoStatus'
      responses:
        '200':
          description: Tasks checklist retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      summary: Create todo item
      description: Registers a new item on the checklist.
      operationId: createTodo
      tags:
        - Todos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /todos/{id}:
    get:
      summary: Get todo details
      description: Retrieve details for a single task.
      operationId: getTodo
      tags:
        - Todos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    patch:
      summary: Update todo details
      description: Mutates target fields (e.g. title, priority, status) of a todo item.
      operationId: updateTodo
      tags:
        - Todos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      summary: Delete todo item
      description: Soft deletes a todo item from the checklist.
      operationId: deleteTodo
      tags:
        - Todos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Task deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived session access token passed in Authorization headers.
    RefreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Secure HTTP-only refresh token used to request new access tokens.
    TelegramSecretVerification:
      type: apiKey
      in: header
      name: X-Telegram-Bot-Api-Secret-Token
      description: Shared bot secret payload validating inbound Webhook calls.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Reserved API key token for external script integrations.
    OAuth2Placeholder:
      type: oauth2
      description: Reserved for future oauth integrations.
      flows:
        authorizationCode:
          authorizationUrl: https://auth.devmate.internal/oauth/authorize
          tokenUrl: https://auth.devmate.internal/oauth/token
          scopes:
            read: Grant view permissions on owned resources.
            write: Grant modification permissions on owned resources.
  parameters:
...
    LimitParam:
      name: limit
      in: query
      description: Max record items count.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CursorParam:
      name: cursor
      in: query
      description: Encoded cursor pointer for page offsets.
      required: false
      schema:
        type: string
...
  headers:
    CorrelationIdResponseHeader:
      description: Ingress trace correlation tracker ID UUID.
      schema:
        type: string
        format: uuid
...
  responses:
    ValidationError:
      description: Input validation constraints violations.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationIdResponseHeader'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorEnvelope'
...
    Unauthorized:
      description: Authentication failure due to missing or expired tokens.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    NotFound:
      description: The requested resource could not be found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    RateLimitError:
      description: Too many requests processed in window.
      headers:
        Retry-After:
          description: Wait seconds before retrying.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ServerError:
      description: Internal system error log tracker.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    UserResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
          required:
            - id
            - telegramId
            - firstName
            - status
          properties:
            id:
              type: string
              format: uuid
            telegramId:
              type: integer
            username:
              type: string
            firstName:
              type: string
            lastName:
              type: string
            status:
              $ref: '#/components/schemas/UserStatus'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    PreferencesResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
          required:
            - baseCurrency
            - timezone
            - language
          properties:
            baseCurrency:
              type: string
              example: USD
            timezone:
              type: string
              example: UTC
            language:
              type: string
              example: en
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    SettingsResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
          required:
            - quietHoursStart
            - quietHoursEnd
            - summaryTime
            - notifyEnabled
          properties:
            quietHoursStart:
              type: string
              example: '22:00'
            quietHoursEnd:
              type: string
              example: '07:00'
            summaryTime:
              type: string
              example: '08:00'
            notifyEnabled:
              type: boolean
              example: true
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    UpdateProfileRequest:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
        username:
          type: string
    UpdatePreferencesRequest:
      type: object
      properties:
        baseCurrency:
          type: string
          maxLength: 3
          minLength: 3
          example: USD
        timezone:
          type: string
          example: Europe/Paris
        language:
          type: string
          maxLength: 5
          example: en
    UpdateSettingsRequest:
      type: object
      properties:
        quietHoursStart:
          type: string
          pattern: '^([0-9]|0[0-9]|1[0-9]|2[0-3]):[0-5][0-9]$'
        quietHoursEnd:
          type: string
          pattern: '^([0-9]|0[0-9]|1[0-9]|2[0-3]):[0-5][0-9]$'
        summaryTime:
          type: string
          pattern: '^([0-9]|0[0-9]|1[0-9]|2[0-3]):[0-5][0-9]$'
        notifyEnabled:
          type: boolean
...
    CreateTodoRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        priority:
          $ref: '#/components/schemas/PriorityLevel'
        dueDate:
          type: string
          format: date-time
    UpdateTodoRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        priority:
          $ref: '#/components/schemas/PriorityLevel'
        status:
          $ref: '#/components/schemas/TodoStatus'
        dueDate:
          type: string
          format: date-time
    TodoResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/TodoDetail'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    TodoDetail:
      type: object
      required:
        - id
        - title
        - priority
        - status
        - version
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        priority:
          $ref: '#/components/schemas/PriorityLevel'
        status:
          $ref: '#/components/schemas/TodoStatus'
        dueDate:
          type: string
          format: date-time
        version:
          type: integer
    TodoListResponse:
      type: object
      required:
        - success
        - data
        - pagination
        - meta
      properties:
        success:
          type: boolean
          const: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/TodoDetail'
        pagination:
          $ref: '#/components/schemas/CursorPaginationMeta'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    ResponseMeta:
      type: object
      required:
        - correlationId
        - traceId
      properties:
        correlationId:
          type: string
          format: uuid
        traceId:
          type: string
          format: uuid
    CursorPaginationMeta:
      type: object
      required:
        - hasMore
      properties:
        nextCursor:
          type: string
        hasMore:
          type: boolean
    ErrorEnvelope:
      type: object
      required:
        - success
        - error
        - meta
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required:
            - code
            - message
            - timestamp
          properties:
            code:
              type: string
              example: RESOURCE_NOT_FOUND
            message:
              type: string
              example: Request resource not found.
            timestamp:
              type: string
              format: date-time
            details:
              type: array
              items:
                type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ValidationErrorEnvelope:
      type: object
      required:
        - success
        - error
        - meta
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required:
            - code
            - message
            - timestamp
            - details
          properties:
            code:
              type: string
              const: VALIDATION_FAILED
            message:
              type: string
              example: Invalid input parameters.
            timestamp:
              type: string
              format: date-time
            details:
              type: array
              items:
                type: object
                required:
                  - field
                  - issue
                properties:
                  field:
                    type: string
                    example: amount
                  issue:
                    type: string
                    example: Value must be a positive string.
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    UserStatus:
      type: string
      enum: [PENDING_ONBOARDING, ACTIVE, SUSPENDED, DEACTIVATED]
    PriorityLevel:
      type: string
      enum: [LOW, MEDIUM, HIGH]
    TodoStatus:
      type: string
      enum: [PENDING, IN_PROGRESS, COMPLETED, OVERDUE]
...